% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mk.proj}
\alias{mk.proj}
\title{Create New Research Project Directory}
\usage{
mk.proj()
}
\description{
create new project directory such as...
 PROJECT
  |- Alias
  |- Analysis
      |- Data
          |- Original
          |- Input
          |- `AvailableData.md`
      |- Figure
      |- Images
  |- Paper
     |- Draft
         |- `Changelog.mdown`
         |- Graphics
         |- `Graphics.mdown`
         |- LaTeX
         |- `README.mdown`
         |- Revise
    |- Submission
  |- Poster
  |- README.mdown
  |- Slide
}
\examples{
mk.proj()
}

